# (POC): Remove some hashmap lookups in distribution lookup