--- a/tpchgen/Cargo.toml
+++ b/tpchgen/Cargo.toml
@@ -8,7 +8,6 @@ authors = ["clflushopt", "alamb"]
 # See ../ARCHITECTURE.md for more details
 [dependencies]
 chrono = "0.4.40"
-indexmap = "2.7.1"
 
 [dev-dependencies]
 flate2 = "1.1.0"