@scsmithr 2025-03-24
> I started going through this PR -- the idea is really nice. I will try and bash out the conversion for the remaining distributions

Yeah my bad for the slowness here. Feel free to take it over.

My only real concern is we lose on flexibility on what distributions we can use. But that's my programmer brain talking -- in reality we only care about that one distribution file.