@alamb 2025-03-24
> Yeah my bad for the slowness here. Feel free to take it over.

No worries -- this is a great step. I also have some ideas about how to avoid hash maps entirely. But I'll do that as a follow on PR potentially. 

Right now me and copilot are going to apply your pattern to all the remaining distributions :)

> My only real concern is we lose on flexibility on what distributions we can use. But that's my programmer brain talking -- in reality we only care about that one distribution file.

Yeah, I agree -- the TPCH data generator hasn't changed for at least 14 years. I don't think flexibility is particularly important